back/requirements_dev.txt
Django==4.0.1
djangorestframework==3.13.1
django-heroku==0.3.1
whitenoise==5.3.0
psycopg2==2.9.3
python-dotenv==0.19.2
django-cors-headers==3.11.0
drf-yasg==1.20.0
